% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-AxisNames.R
\name{add_relative_pos}
\alias{add_relative_pos}
\alias{add_rel_pos}
\title{Given a flat \code{\link[=AxisNames]{AxisNames()}} object, add relative positions
to each \code{ConstantAstNode} in the list. This is used for
comparing \code{ConstantAstNode} objects in the AST, as they may
have the same count, different/same starts, but same/different
relative positions.}
\usage{
add_relative_pos(axes)

add_rel_pos(axes)
}
\arguments{
\item{axes}{a flat \code{\link[=AxisNames]{AxisNames()}} object}
}
\value{
a modified \code{\link[=AxisNames]{AxisNames()}} object with \code{relative_pos} added to
each \code{ConstantAstNode} in the list.
}
\description{
Given a flat \code{\link[=AxisNames]{AxisNames()}} object, add relative positions
to each \code{ConstantAstNode} in the list. This is used for
comparing \code{ConstantAstNode} objects in the AST, as they may
have the same count, different/same starts, but same/different
relative positions.
}
\keyword{internal}
