% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-semantic-analyzer.R
\name{analyse_semantics}
\alias{analyse_semantics}
\title{Semantic Analyzer for Einops AST}
\usage{
analyse_semantics(
  ast,
  op_kind = c("rearrange", "reduce", "repeat"),
  axes_lengths_hint = list()
)
}
\arguments{
\item{ast}{EinopsAst object}

\item{op_kind}{One of 'rearrange', 'reduce', 'repeat'}

\item{axes_lengths_hint}{Named list of axis lengths (optional)}
}
\value{
SemanticInfo S3 object
}
\description{
Analyze an EinopsAst and return a pure SemanticInfo S3 object
}
\keyword{internal}
