% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-apply_recipe.R
\name{apply_recipe}
\alias{apply_recipe}
\title{Apply a TransformRecipe to a Tensor}
\usage{
apply_recipe(backend, recipe, tensor, reduction_type, axes_lengths)
}
\arguments{
\item{backend}{The \code{\link[=EinopsBackend]{EinopsBackend()}} to use for tensor operations.}

\item{recipe}{A \code{\link[=TransformRecipe]{TransformRecipe()}} that specifies how to transform the
tensor.}

\item{tensor}{A tensor to be transformed}

\item{reduction_type}{A character of length 1 that specifies the type of
reduction to apply.}

\item{axes_lengths}{TODO check - I think its just an integer vector?}
}
\value{
A tensor that has been transformed according to the recipe, with the
same type (if possible) as the input tensor.
}
\description{
This function applies a TransformRecipe to a tensor, performing rearrangement
and reduction as specified by the recipe.

TODO docs for which parts of the compilation pipeline this is.
}
\keyword{internal}
