% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{as.character,EinopsAst-method}
\alias{as.character,EinopsAst-method}
\title{Convert EinopsAst back to string representation}
\usage{
\S4method{as.character}{EinopsAst}(x)
}
\arguments{
\item{x}{EinopsAst object}
}
\value{
character string canonical representation
}
\description{
Convert EinopsAst back to string representation
}
