% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-AxisNames.R
\name{as_iterables}
\alias{as_iterables}
\title{Convert an object to a list of iterables (lists)}
\usage{
as_iterables(x, ...)
}
\arguments{
\item{x}{the input object to be converted}

\item{...}{additional arguments (not used)}
}
\value{
a list of iterables, where each iterable is a list of elements
}
\description{
This function converts an object into a list of iterables, where each
iterable is a list of elements. The function is generic and can be
extended for different classes.
}
\section{AxisNames}{

This method converts an \code{AxisNames} object into a list of iterables.
Each element in the \code{AxisNames} is processed to extract its name or
constant value. If the element is a \code{ConstantAstNode} with a count of
1, it is converted to an empty list. Otherwise, it is wrapped in
a list. If the element is a \code{NamedAxisAstNode}, its name is extracted.
Regular AxisNames that are nested will be unclassed
}

\keyword{internal}
