% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{ast_to_tokens}
\alias{ast_to_tokens}
\title{Convert EinopsAst back to TokenSequence}
\usage{
ast_to_tokens(ast)
}
\arguments{
\item{ast}{EinopsAst object to convert}
}
\value{
EinopsTokenSequence object
}
\description{
Converts an EinopsAst object back to a TokenSequence that represents the original pattern
}
\keyword{internal}
