% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{AsteriskAstNode}
\alias{AsteriskAstNode}
\title{Create an AsteriskAstNode}
\usage{
AsteriskAstNode(src = list())
}
\arguments{
\item{src}{List with start position}
}
\value{
AsteriskAstNode object
}
\description{
Create an AsteriskAstNode
}
\keyword{internal}
