% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{axes_to_tokens}
\alias{axes_to_tokens}
\title{Convert list of axis nodes to tokens}
\usage{
axes_to_tokens(axes)
}
\arguments{
\item{axes}{List of AST axis nodes}
}
\value{
List of tokens
}
\description{
Helper function to convert a list of AST axis nodes to tokens
}
\keyword{internal}
