% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{axis_to_tokens}
\alias{axis_to_tokens}
\title{Convert single axis node to tokens}
\usage{
axis_to_tokens(axis, start_pos)
}
\arguments{
\item{axis}{Single AST axis node}

\item{start_pos}{Starting position for the first token}
}
\value{
List of tokens representing this axis
}
\description{
Helper function to convert a single AST axis node to tokens
}
\keyword{internal}
