% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-AxisNames.R
\name{AxisNames}
\alias{AxisNames}
\title{Construct an instance of an \code{AxisNames} class}
\usage{
AxisNames(...)
}
\arguments{
\item{...}{a list of elements or arbitrary number of elements}
}
\value{
an \code{AxisNames} object, which is a list with a specific class
}
\description{
This is a wrapper for a \code{\link[=list]{list()}}, but the elements may only be singular
\code{\link[=character]{character()}}, \code{\link[=ConstantAstNode]{ConstantAstNode()}} or other AstNode objects. (only 1 level
of nesting allowed, as its essentially another representation of a
\code{\link[=OneSidedAstNode]{OneSidedAstNode()}}.

Note that when using \code{\link[=c]{c()}} on an \code{AxisNames} object, when the other object is
another \code{\link[=AxisNames]{AxisNames()}}, the elements of that will simply be appended to the
first object. To nest (in the case of a \code{GroupAstNode}), you must append
a \code{list(AxisNames(...))} to the first object instead.
}
\keyword{internal}
