% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calculate_output_dimensions}
\alias{calculate_output_dimensions}
\title{Calculate output dimensions from pattern}
\usage{
calculate_output_dimensions(pattern, current_dims)
}
\arguments{
\item{pattern}{output pattern list}

\item{current_dims}{current array dimensions}
}
\value{
integer vector of new dimensions
}
\description{
Calculate output dimensions from pattern
}
\keyword{internal}
