% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/utils.R
\name{count_non_ellipsis_axes}
\alias{count_non_ellipsis_axes}
\title{Count non-ellipsis axes in a pattern}
\usage{
count_non_ellipsis_axes(pattern)

count_non_ellipsis_axes(pattern)
}
\arguments{
\item{pattern}{list representing parsed pattern}
}
\value{
integer count of non-ellipsis axes

integer count of non-ellipsis axes
}
\description{
Count non-ellipsis axes in a pattern

Count non-ellipsis axes in a pattern
}
\keyword{internal}
