% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/utils.R
\name{create_dimension_mapping}
\alias{create_dimension_mapping}
\title{Create mapping from axes to array dimensions}
\usage{
create_dimension_mapping(pattern, dims, ellipsis_count)

create_dimension_mapping(pattern, dims, ellipsis_count)
}
\arguments{
\item{pattern}{list representing input pattern}

\item{dims}{integer vector of array dimensions}

\item{ellipsis_count}{integer number of dimensions represented by ellipsis}
}
\value{
list with axis mappings

list with axis mappings
}
\description{
Create mapping from axes to array dimensions

Create mapping from axes to array dimensions
}
\keyword{internal}
