% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-apply_recipe.R
\name{create_execution_plan}
\alias{create_execution_plan}
\title{Reconstruct all actual parameters using shape.}
\usage{
create_execution_plan(recipe, shape, axes_dims)
}
\arguments{
\item{recipe}{a populated \code{\link[=TransformRecipe]{TransformRecipe()}} object}

\item{shape}{A vector of integers representing the shape of the tensor.}

\item{axes_dims}{A named list of axes names to their dimensions/lengths.}
}
\value{
An \code{EinopsExecutionPlan} object that contains the execution plan for
transforming tensors according to the specified recipe and shape.
}
\description{
This implements \verb{_reconstruct_from_shape_uncached} from the python
implementation of einops. In the future, a cached version may be
implemented.
}
\keyword{internal}
