% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexer.R, R/parser.R
\name{.next_token}
\alias{.next_token}
\title{.next_token}
\usage{
.next_token(tokens, pos)

.next_token(tokens, pos)
}
\arguments{
\item{tokens}{list of tokens}

\item{pos}{current position}
}
\value{
list with token and new position

list with token and new position
}
\description{
Internal helper to get next token from stream

Internal helper to get next token from stream
}
\keyword{internal}
