% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/einops-package.R
\docType{package}
\name{einops-package}
\alias{einops}
\alias{einops-package}
\title{einops: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Qile0317/einops}
  \item Report bugs at \url{https://github.com/Qile0317/einops/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
