% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\docType{class}
\name{EinopsAst-class}
\alias{EinopsAst-class}
\title{S4 class for Einops Abstract Syntax Tree}
\description{
S4 class for Einops Abstract Syntax Tree
}
\section{Slots}{

\describe{
\item{\code{text}}{character string of original pattern}

\item{\code{axes}}{list of axis descriptors}

\item{\code{ellipsis_pos}}{integer position of ellipsis or NA}
}}

