% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-constraint-checker.R
\name{EinopsConstraint}
\alias{EinopsConstraint}
\title{Einops Constraint Base Class}
\usage{
EinopsConstraint()
}
\description{
Base class for operation-specific constraints
}
\keyword{internal}
