% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-constraint-checker.R
\name{EinopsRearrangeConstraint}
\alias{EinopsRearrangeConstraint}
\title{Einops Rearrange Constraint}
\usage{
EinopsRearrangeConstraint()
}
\description{
Constraint object for rearrange operation
}
\keyword{internal}
