% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-constraint-checker.R
\name{EinopsReduceConstraint}
\alias{EinopsReduceConstraint}
\title{Einops Reduce Constraint}
\usage{
EinopsReduceConstraint()
}
\description{
Constraint object for reduce operation
}
\keyword{internal}
