% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-constraint-checker.R
\name{EinopsRepeatConstraint}
\alias{EinopsRepeatConstraint}
\title{Einops Repeat Constraint}
\usage{
EinopsRepeatConstraint()
}
\description{
Constraint object for repeat operation
}
\keyword{internal}
