% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-token.R
\name{EinopsTokenSequence}
\alias{EinopsTokenSequence}
\title{TokenSequence}
\usage{
EinopsTokenSequence(...)
}
\arguments{
\item{...}{tokens or EinopsTokenSequences to include}
}
\value{
list of tokens
}
\description{
Helper to build a token sequence (list of tokens)
}
\keyword{internal}
