% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_nodes.R
\name{EllipsisAstNode}
\alias{EllipsisAstNode}
\title{Create an EllipsisAstNode}
\usage{
EllipsisAstNode(src)
}
\arguments{
\item{src}{List with start and end positions}
}
\value{
EllipsisAstNode object
}
\description{
Create an EllipsisAstNode
}
\keyword{internal}
