% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-TransformRecipe.R
\name{expand_ellipsis}
\alias{expand_ellipsis}
\title{Expand ellipses of an EinopsAst}
\usage{
expand_ellipsis(einops_ast, ndim)
}
\arguments{
\item{einops_ast}{an EinopsAst}

\item{ndim}{integer. Number of dimensions in the input tensor}
}
\value{
an expanded EinopsAst with ellipses expanded
}
\description{
Helper for \code{\link[=prepare_transformation_recipe]{prepare_transformation_recipe()}}.

This function expands each relevant ellipsis ast node inplace into a
sequence of \code{NamedAxisAstNode} nodes, where each node will have
a name like \code{"...1"}, \code{"...2"}, etc. and an empty \code{src} list.

Also does some further validation of the ellipses syntax using \code{ndim}.
}
\keyword{internal}
