% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{extract_axis_names}
\alias{extract_axis_names}
\title{Extract axis names from a parsed pattern}
\usage{
extract_axis_names(pattern)
}
\arguments{
\item{pattern}{list representing parsed pattern}
}
\value{
character vector of axis names
}
\description{
Extract axis names from a parsed pattern
}
\keyword{internal}
