% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{extract_axis_names_flat}
\alias{extract_axis_names_flat}
\title{Extract axis names from pattern (flattened)}
\usage{
extract_axis_names_flat(pattern)
}
\arguments{
\item{pattern}{list representing pattern}
}
\value{
character vector of axis names
}
\description{
Extract axis names from pattern (flattened)
}
\keyword{internal}
