% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{find_ellipsis_axes}
\alias{find_ellipsis_axes}
\title{Check for ellipsis in pattern}
\usage{
find_ellipsis_axes(pattern)
}
\arguments{
\item{pattern}{list representing parsed pattern}
}
\value{
character vector of axes that contain ellipsis
}
\description{
Check for ellipsis in pattern
}
\keyword{internal}
