% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_einops_ast.R
\name{find_top_level_arrow}
\alias{find_top_level_arrow}
\title{Einops AST Parser}
\usage{
find_top_level_arrow(tokens)
}
\arguments{
\item{tokens}{EinopsTokenSequence object}
}
\value{
Integer position of the top-level arrow token
}
\description{
Iterative parser for einops patterns that converts token sequences to AST
}
\keyword{internal}
