% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-parser.R
\name{find_top_level_arrow_index}
\alias{find_top_level_arrow_index}
\title{Find the arrow in a token sequence}
\usage{
find_top_level_arrow_index(tokens)
}
\arguments{
\item{tokens}{EinopsTokenSequence object}
}
\value{
one-indexed Integer position of the arrow token
}
\description{
Find the arrow in a token sequence
}
\keyword{internal}
