% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-parser.R
\name{generate_constructor}
\alias{generate_constructor}
\title{Generate constructor code for AST node}
\usage{
generate_constructor(node)
}
\arguments{
\item{node}{AST node}
}
\value{
Character string with constructor call
}
\description{
Generate constructor code for AST node
}
\keyword{internal}
