% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-backends.R
\name{get_backend_registry}
\alias{get_backend_registry}
\title{Get the singleton backend registry instance}
\usage{
get_backend_registry(
  clear_testing = !identical(Sys.getenv("TESTTHAT"), "true")
)
}
\arguments{
\item{refresh}{Logical flag indicating whether to refresh the registry.}
}
\value{
A singleton instance of \code{\link[=BackendRegistry]{BackendRegistry()}}.
}
\description{
This function retrieves the singleton instance of the backend registry.
If the instance does not exist, it creates a new one.
If \code{refresh} is set to TRUE, it will create a new instance even if
one already exists.
}
\keyword{internal}
