% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-AxisNames.R
\name{get_identifiers}
\alias{get_identifiers}
\title{Given a \code{OneSidedAstNode} object, get unique identifiers}
\usage{
get_identifiers(ast, add_relative_pos = FALSE, ...)
}
\arguments{
\item{ast}{the Abstract Syntax Tree (AST) of the einops expression}

\item{add_relative_pos}{a boolean indicating whether to add relative
positions to the \code{ConstantAstNode} objects in the output.}

\item{...}{additional arguments (not used)}
}
\value{
an \code{\link[=AxisNames]{AxisNames()}} of unique identifiers
}
\description{
get an unordered list, representing a set of all unique
identifiers on one side of the expression. Named nodes become
the character representing its name, ConstantAstNode nodes
are just themselves, aside from 1 which is ignored. GroupAstNode
nodes are flattened, and EllipsisAstNode nodes are ignored.
}
\keyword{internal}
