% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-AxisNames.R
\name{get_ordered_axis_names}
\alias{get_ordered_axis_names}
\title{Get the ordered axis names from a OneSidedAstNode, removing
all nesting.}
\usage{
get_ordered_axis_names(ast, ...)
}
\arguments{
\item{ast}{OneSidedAstNode object}

\item{...}{additional arguments (not used)}
}
\value{
a \code{\link[=list]{list()}} of axis names
}
\description{
Get the ordered axis names from a OneSidedAstNode, removing
all nesting.
}
\keyword{internal}
