% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-AxisNames.R
\name{get_reduced_axis_names}
\alias{get_reduced_axis_names}
\title{Get reduced axis names by removing axes present in y from x}
\usage{
get_reduced_axis_names(x, y, ...)
}
\arguments{
\item{x}{\code{\link[=AxisNames]{AxisNames()}} object to reduce}

\item{y}{\code{\link[=AxisNames]{AxisNames()}} object containing axes to remove}

\item{...}{additional arguments (not used)}
}
\value{
AxisNames object with axes from x that are not in y.
Note that the only element left in each constant node's src
list will be \code{relative_pos}.
}
\description{
Get reduced axis names by removing axes present in y from x
}
\keyword{internal}
