% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{get_ungrouped_nodes}
\alias{get_ungrouped_nodes}
\title{get flat list of all AstNodes in order, including children of
GroupAstNodes.}
\usage{
get_ungrouped_nodes(ast, ...)
}
\arguments{
\item{ast}{An AstNode object}
}
\value{
A flat \code{list} of AstNodes
}
\description{
get flat list of all AstNodes in order, including children of
GroupAstNodes.
}
\keyword{internal}
