% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_nodes.R
\name{GroupAstNode}
\alias{GroupAstNode}
\title{Create a GroupAstNode}
\usage{
GroupAstNode(children, src)
}
\arguments{
\item{children}{List of axis nodes contained in this group}

\item{src}{List with start and end positions}
}
\value{
GroupAstNode object
}
\description{
Create a GroupAstNode
}
\keyword{internal}
