% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexer.R
\name{lex}
\alias{lex}
\title{Lexically analyze einops pattern into structured tokens}
\usage{
lex(pattern)
}
\arguments{
\item{pattern}{character string with einops pattern}
}
\value{
list of token objects with type, value, start, end fields
}
\description{
Lexically analyze einops pattern into structured tokens
}
\keyword{internal}
