% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{make_ast}
\alias{make_ast}
\title{Make AST from parsed terms}
\usage{
make_ast(terms, pattern, op)
}
\arguments{
\item{terms}{list of term objects from parser}

\item{pattern}{original pattern string}

\item{op}{reduction operation}
}
\value{
EinopsAst object
}
\description{
Make AST from parsed terms
}
