% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_shape.R
\name{match_dimensions}
\alias{match_dimensions}
\title{Match tokens to their dimensional values}
\usage{
match_dimensions(tokens, dimensions)
}
\arguments{
\item{tokens}{EinopsTokenSequence}

\item{dimensions}{numeric vector}
}
\value{
named list mapping dimension names to values
}
\description{
Match tokens to their dimensional values
}
\keyword{internal}
