% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast_nodes.R
\name{merge_src}
\alias{merge_src}
\title{Merge source position information}
\usage{
merge_src(src_a, src_b)
}
\arguments{
\item{src_a}{First source position (list with start, end)}

\item{src_b}{Second source position (list with start, end)}
}
\value{
Combined source position with earliest start and latest end
}
\description{
Merge source position information
}
\keyword{internal}
