% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{NamedAxisAstNode}
\alias{NamedAxisAstNode}
\title{Create a NamedAxisAstNode}
\usage{
NamedAxisAstNode(name, src = list())
}
\arguments{
\item{name}{Character string, the name of the axis}

\item{src}{List with start position}
}
\value{
NamedAxisAstNode object
}
\description{
Create a NamedAxisAstNode
}
\keyword{internal}
