% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{NothingAstNode}
\alias{NothingAstNode}
\title{Create a NothingAstNode}
\usage{
NothingAstNode()
}
\arguments{
\item{src}{List with start position}
}
\value{
NothingAstNode object
}
\description{
Create a NothingAstNode
}
\keyword{internal}
