% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{OneSidedAstNode}
\alias{OneSidedAstNode}
\title{Create a OneSidedAstNode (wrapper for input/output axes lists)}
\usage{
OneSidedAstNode(...)
}
\arguments{
\item{...}{Axis nodes (or a single list of axis nodes) If this is already
a OneSidedAstNode, it is returned as is.}
}
\value{
OneSidedAstNode object
}
\description{
Create a OneSidedAstNode (wrapper for input/output axes lists)
}
\keyword{internal}
