% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_einops_ast.R
\name{parse_axes_iter}
\alias{parse_axes_iter}
\title{Parse a sequence of axis tokens iteratively}
\usage{
parse_axes_iter(tokens)
}
\arguments{
\item{tokens}{List of tokens representing one side of the pattern}
}
\value{
List of AST nodes
}
\description{
Parse a sequence of axis tokens iteratively
}
\keyword{internal}
