% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-parser.R
\name{parse_einops_ast}
\alias{parse_einops_ast}
\title{Parse einops pattern into AST}
\usage{
parse_einops_ast(tokens)
}
\arguments{
\item{tokens}{EinopsTokenSequence object from the lexer}
}
\value{
EinopsAst object
}
\description{
In the python implementation, they preprocess all 1's
into () and then parse it into the equivalent of a
GroupAstNode. Here, we won't do that
}
\keyword{internal}
