% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-parser.R
\name{parse_onesided_ast}
\alias{parse_onesided_ast}
\title{Parse a one-sided einops pattern into OneSidedAstNode}
\usage{
parse_onesided_ast(tokens)
}
\arguments{
\item{tokens}{EinopsTokenSequence object from the lexer (no arrow expected)}
}
\value{
OneSidedAstNode object
}
\description{
Parse a one-sided einops pattern into OneSidedAstNode
}
\keyword{internal}
