% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_shape.R
\name{parse_shape}
\alias{parse_shape}
\title{Parse shape expression}
\usage{
parse_shape(x, expr, ...)
}
\arguments{
\item{expr}{expression to parse}
}
\value{
named list of tokens to their mapped dimensions
}
\description{
This function is the entry point for parsing a shape expression.
}
