% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{parse_tokens}
\alias{parse_tokens}
\title{Parse tokens into a nested structure}
\usage{
parse_tokens(tokens)
}
\arguments{
\item{tokens}{list of tokens}
}
\value{
list representing the parsed structure
}
\description{
Parse tokens into a nested structure
}
\keyword{internal}
