% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-repr.R
\name{pprint}
\alias{pprint}
\title{Pretty Print}
\usage{
pprint(x, ...)
}
\arguments{
\item{x}{Object to pretty print}

\item{...}{Additional arguments passed to \code{repr()}}
}
\description{
This is a convenience function that prints the \code{\link[=repr]{repr()}} of an object.
It is similar to python's \code{pprint.pprint()}.
}
\keyword{internal}
