% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{print.AstNode}
\alias{print.AstNode}
\title{Print method for AstNode}
\usage{
\method{print}{AstNode}(x, ...)
}
\arguments{
\item{x}{AstNode object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for AstNode
}
