% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{print.EinopsAst}
\alias{print.EinopsAst}
\title{Print method for EinopsAst}
\usage{
\method{print}{EinopsAst}(x, ...)
}
\arguments{
\item{x}{EinopsAst object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for EinopsAst
}
