% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-token.R
\name{print.EinopsToken}
\alias{print.EinopsToken}
\title{Print method for EinopsToken}
\usage{
\method{print}{EinopsToken}(x, ...)
}
\arguments{
\item{x}{EinopsToken object}

\item{...}{additional arguments (unused)}
}
\value{
invisible x
}
\description{
Print EinopsToken objects in a clean format showing construction
}
