% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-token.R
\name{print.EinopsTokenSequence}
\alias{print.EinopsTokenSequence}
\title{Print method for EinopsTokenSequence
Print EinopsTokenSequences w/reconstructed expression & constructor}
\usage{
\method{print}{EinopsTokenSequence}(x, ...)
}
\arguments{
\item{x}{EinopsTokenSequence object}

\item{...}{additional arguments (unused)}
}
\value{
invisible x
}
\description{
Print method for EinopsTokenSequence
Print EinopsTokenSequences w/reconstructed expression & constructor
}
