% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-parser.R
\name{reconstruct_node}
\alias{reconstruct_node}
\title{Reconstruct expression text from AST node}
\usage{
reconstruct_node(node)
}
\arguments{
\item{node}{AST node (NamedAxisAstNode, EllipsisAstNode, or GroupAstNode)}
}
\value{
Character string representation
}
\description{
Reconstruct expression text from AST node
}
\keyword{internal}
