% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-repr.R
\name{repr}
\alias{repr}
\title{Create a string representation of an object, as a vector of lines}
\usage{
repr(x, indent = 0L, ...)
}
\arguments{
\item{x}{Object to represent}

\item{indent}{Indentation level (number of spaces)}

\item{...}{Additional arguments passed to methods}
}
\value{
A character vector of class c("repr_output", "character"), each element is a line
}
\description{
Create a string representation of an object, as a vector of lines
}
\keyword{internal}
