% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/utils.R
\name{reshape_result}
\alias{reshape_result}
\title{Reshape result according to output pattern}
\usage{
reshape_result(result, reduction_plan)

reshape_result(result, reduction_plan)
}
\arguments{
\item{result}{reduced array}

\item{reduction_plan}{list from create_reduction_plan}
}
\value{
reshaped array

reshaped array
}
\description{
Reshape result according to output pattern

Reshape result according to output pattern
}
\keyword{internal}
