% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{show,EinopsAst-method}
\alias{show,EinopsAst-method}
\title{Print method for EinopsAst}
\usage{
\S4method{show}{EinopsAst}(object)
}
\arguments{
\item{x}{EinopsAst object}
}
\description{
Print method for EinopsAst
}
\keyword{internal}
