% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{tail.GroupAstNode}
\alias{tail.GroupAstNode}
\title{Get the last n children of a GroupAstNode as a list of AstNodes.}
\usage{
\method{tail}{GroupAstNode}(x, n = 1)
}
\description{
Get the last n children of a GroupAstNode as a list of AstNodes.
}
\keyword{internal}
