% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-backends.R
\name{thunk}
\alias{thunk}
\title{Simple thunk: wraps an input in a no-argument function}
\usage{
thunk(input)
}
\arguments{
\item{input}{Any R object or expression}
}
\value{
A thunk that returns the object when called with
class \code{c("thunk", "function")}
}
\description{
Simple thunk: wraps an input in a no-argument function
}
\keyword{internal}
