% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{tokenize_einops}
\alias{tokenize_einops}
\title{Tokenize an einops expression string}
\usage{
tokenize_einops(expr)
}
\arguments{
\item{expr}{character string containing the einops expression}
}
\value{
list of tokens
}
\description{
Tokenize an einops expression string
}
\keyword{internal}
