% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexer.R
\name{TokenSequence}
\alias{TokenSequence}
\title{TokenSequence}
\usage{
TokenSequence(...)
}
\arguments{
\item{...}{tokens to include}
}
\value{
list of tokens
}
\description{
Helper to build a token sequence (list of tokens)
}
\keyword{internal}
