% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-TransformRecipe.R
\name{TransformRecipe}
\alias{TransformRecipe}
\title{TransformRecipe S3 constructor}
\usage{
TransformRecipe(
  elementary_axes_lengths,
  axis_name2elementary_axis,
  input_composition_known_unknown,
  axes_permutation,
  first_reduced_axis,
  added_axes,
  output_composite_axes
)
}
\arguments{
\item{elementary_axes_lengths}{Integer vector. List of sizes for elementary axes as they appear in left expression.}

\item{axis_name2elementary_axis}{Named integer vector. Mapping from name to position.}

\item{input_composition_known_unknown}{List of list(present, unknown) integer vectors. Each element is a tuple of known and unknown indices.}

\item{axes_permutation}{Integer vector. Permutation applied to elementary axes.}

\item{first_reduced_axis}{Integer. First position of reduced axes.}

\item{added_axes}{Named integer vector. Axis position -> axis index.}

\item{output_composite_axes}{List of integer vectors. Ids of axes as they appear in result.}
}
\value{
An object of class 'TransformRecipe'.
}
\description{
Recipe describes actual computation pathway. Can be applied to a tensor or variable.
}
