% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{UnderscoreAstNode}
\alias{UnderscoreAstNode}
\title{Create an UnderscoreAstNode}
\usage{
UnderscoreAstNode(src = list())
}
\arguments{
\item{src}{List with start position}
}
\value{
UnderscoreAstNode object
}
\description{
Create an UnderscoreAstNode
}
\keyword{internal}
