% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing.R
\name{unpack}
\alias{unpack}
\title{Unpack a single tensor to several in a list}
\usage{
unpack(tensor, packed_shapes, pattern)
}
\arguments{
\item{tensor}{tensor to be unpacked}

\item{packed_shapes}{list of shapes that take the place of \code{'*'} in each
output. will contain a single tensor for every provided shape}

\item{pattern}{a string pattern that is shared for input and all outputs,
e.g. \code{"i j * k"} or \code{"batch seq *"}, where * designates an axis to be
unpacked.}
}
\value{
\code{\link[=list]{list()}} of tensors of the same type
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Unpacks a single tensor into several by splitting over a selected axes.
See einops tutorial for introduction into packing (and how it replaces
stack and concatenation).

If framework supports views, results are views to the original tensor.
}
\examples{
library(zeallot)
inputs <- list(
    array(0, c(2, 3, 5)),
    array(0, c(2, 3, 7, 5)),
    array(0, c(2, 3, 7, 9, 5))
)
packed, ps \%<-\% pack(inputs, 'i j * k')
# the line above is identical to assigning each of those
# variables to the first and second elements of the output

print(dim(packed))
#> c(2, 3, 71, 5)
print(ps)
#> list(integer(), 7, c(7, 9))

# In this example, axes were matched to: i=2, j=3, k=5 based on order
# (first, second, and last). All other axes were 'packed' and
# concatenated. PS (packed shapes) contains information about axes
# matched to '*' in every input. The resulting tensor has as many
# elements as all inputs in total.

# Packing can be reversed with unpack, which additionally needs PS
# (packed shapes) to reconstruct order.

# inputs_unpacked <- unpack(packed, ps, 'i j * k')
# [x.shape for x in inputs_unpacked]
# [(2, 3, 5), (2, 3, 7, 5), (2, 3, 7, 9, 5)]

# Read the tutorial for introduction and application
}
\seealso{
\code{\link[=pack]{pack()}}
}
\keyword{internal}
