% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_shape.R
\name{validate_dimension_count}
\alias{validate_dimension_count}
\title{Validate dimension count matches token count}
\usage{
validate_dimension_count(tokens, dimensions)
}
\arguments{
\item{tokens}{EinopsTokenSequence}

\item{dimensions}{numeric vector}
}
\description{
Validate dimension count matches token count
}
\keyword{internal}
