% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-semantic-validator.R
\name{validate_reduction_operation}
\alias{validate_reduction_operation}
\title{Syntactically Validate the reduction operation in the einops
expression}
\usage{
validate_reduction_operation(einops_ast, operation)
}
\arguments{
\item{einops_ast}{the parsed Abstract Syntax Tree (AST) of the einops
expression}

\item{operation}{the reduction operation, either a string or a function}
}
\value{
the input einops_ast (invisibly) if the validation passes
}
\description{
Syntactically Validate the reduction operation in the einops
expression
}
\keyword{internal}
