% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_shape.R
\name{validate_syntax}
\alias{validate_syntax}
\title{Validate token syntax}
\usage{
validate_syntax(tokens)
}
\arguments{
\item{tokens}{EinopsTokenSequence}
}
\description{
Validate token syntax
}
\keyword{internal}
