% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-AxisNames.R
\name{add_relative_pos}
\alias{add_relative_pos}
\title{Given a flat \code{\link[=AxisNames]{AxisNames()}} object, add relative positions
to each \code{ConstantAstNode} in the list. This is used for
comparing \code{ConstantAstNode} objects in the AST, as they may
have the same count, different/same starts, but same/different
relative positions.}
\usage{
add_relative_pos(axes, rm_other_src_elements = FALSE, ...)
}
\arguments{
\item{axes}{a flat \code{\link[=AxisNames]{AxisNames()}} object}

\item{rm_other_src_elements}{a boolean indicating whether to remove
other source elements from the \code{ConstantAstNode} objects. If TRUE,
the only element within the \code{src} list of each \code{ConstantAstNode}
will be \code{relative_pos}, and all other elements will be removed.}
}
\value{
a modified \code{\link[=AxisNames]{AxisNames()}} object with \code{relative_pos} added to
each \code{ConstantAstNode} in the list.
}
\description{
Given a flat \code{\link[=AxisNames]{AxisNames()}} object, add relative positions
to each \code{ConstantAstNode} in the list. This is used for
comparing \code{ConstantAstNode} objects in the AST, as they may
have the same count, different/same starts, but same/different
relative positions.
}
\keyword{internal}
