% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-AxisNames.R
\name{as_axis_names}
\alias{as_axis_names}
\title{Convert an AstNode into an \code{\link[=AxisNames]{AxisNames()}} object}
\usage{
as_axis_names(ast, ...)
}
\arguments{
\item{ast}{the AstNode input}

\item{...}{other args}
}
\value{
an \code{\link[=AxisNames]{AxisNames()}} object of the axes in order
}
\description{
Convert an AstNode into an \code{\link[=AxisNames]{AxisNames()}} object
}
\examples{
get_axis_names(parse_onesided_ast(lex('a (b c 1) 2 d')))
# should output AxisNames(
    "a", AxisNames("b", "c", ConstantAstNode(1)), ConstantAstNode(2), "d"
)
}
\keyword{internal}
