% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-backends.R
\name{BackendRegistry}
\alias{BackendRegistry}
\title{Singleton Backend Registry, managing all available backends.}
\description{
Contains global backend pool
}
\keyword{internal}
\section{Super class}{
\code{\link[R6P:Singleton]{R6P::Singleton}} -> \code{BackendRegistry}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BackendRegistry-get_backend}{\code{BackendRegistry$get_backend()}}
\item \href{#method-BackendRegistry-set_debug}{\code{BackendRegistry$set_debug()}}
\item \href{#method-BackendRegistry-register_backend}{\code{BackendRegistry$register_backend()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="R6P" data-topic="Singleton" data-id="initialize"><a href='../../R6P/html/Singleton.html#method-Singleton-initialize'><code>R6P::Singleton$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BackendRegistry-get_backend"></a>}}
\if{latex}{\out{\hypertarget{method-BackendRegistry-get_backend}{}}}
\subsection{Method \code{get_backend()}}{
detect the return relevant backend from the input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackendRegistry$get_backend(tensor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tensor}}{any support tensor-like class}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A singleton instance of a \code{\link[=BackendRegistry]{BackendRegistry()}} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BackendRegistry-set_debug"></a>}}
\if{latex}{\out{\hypertarget{method-BackendRegistry-set_debug}{}}}
\subsection{Method \code{set_debug()}}{
Set whether debug messages should be displayed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackendRegistry$set_debug(flag = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flag}}{boolean}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
this object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BackendRegistry-register_backend"></a>}}
\if{latex}{\out{\hypertarget{method-BackendRegistry-register_backend}{}}}
\subsection{Method \code{register_backend()}}{
Register a new backend singleton
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BackendRegistry$register_backend(backend_class)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{backend_class}}{an EinopsBackend subclass generator}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
this object
}
}
}
