% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{ConstantAstNode}
\alias{ConstantAstNode}
\title{Create a ConstantAstNode}
\usage{
ConstantAstNode(count, src)
}
\arguments{
\item{count}{Character string representing the constant value}

\item{src}{List with start position}
}
\value{
ConstantAstNode object
}
\description{
Create a ConstantAstNode
}
\keyword{internal}
