% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/einops-package.R
\docType{package}
\name{einops-package}
\alias{einops}
\alias{einops-package}
\title{einops: Flexible, Powerful, and Readable Tensor Operations}
\description{
Perform tensor operations using a concise yet expressive syntax inspired by the Python library of the same name. Reshape, rearrange, and combine multidimensional arrays for scientific computing, machine learning, and data analysis. Einops simplifies complex manipulations, making code more maintainable and intuitive. The original implementation is demonstrated in Rogozhnikov (2022) \url{https://openreview.net/forum?id=oapKSVM2bcj}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Qile0317/einops}
  \item \url{https://qile0317.github.io/einops/}
  \item Report bugs at \url{https://github.com/Qile0317/einops/issues}
}

}
\author{
\strong{Maintainer}: Qile Yang \email{qile.yang@berkeley.edu} (\href{https://orcid.org/0009-0005-0148-2499}{ORCID}) [copyright holder]

}
\keyword{internal}
