% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/einops-package.R
\docType{package}
\name{einops-package}
\alias{einops}
\alias{einops-package}
\title{einops: Flexible and powerful tensor operations for readable and reliable code}
\description{
R port of the einops library for flexible and powerful array and tensor operations. It provides a concise and readable syntax for reshaping, rearranging, and combination.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Qile0317/einops}
  \item \url{https://qile0317.github.io/einops/}
  \item Report bugs at \url{https://github.com/Qile0317/einops/issues}
}

}
\author{
\strong{Maintainer}: Qile Yang \email{qile.yang@berkeley.edu} (\href{https://orcid.org/0009-0005-0148-2499}{ORCID}) [copyright holder]

}
\keyword{internal}
