% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{EinopsAst}
\alias{EinopsAst}
\title{Create an EinopsAst root node}
\usage{
EinopsAst(input_axes, output_axes, src)
}
\arguments{
\item{input_axes}{List of axis nodes for the input pattern, or a NothingAstNode, or OneSidedAstNode}

\item{output_axes}{List of axis nodes for the input pattern, or a NothingAstNode, or OneSidedAstNode}

\item{src}{List with start position covering the full pattern}
}
\value{
EinopsAst object
}
\description{
Create an EinopsAst root node
}
\keyword{internal}
