% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_shape.R
\name{EinopsShapeSemanticAst}
\alias{EinopsShapeSemanticAst}
\title{EinopsShapeSemanticAst constructor}
\usage{
EinopsShapeSemanticAst(tokens, dimensions)
}
\arguments{
\item{tokens}{parsed EinopsTokenSequence}

\item{dimensions}{numeric vector of array dimensions}
}
\value{
EinopsShapeSemanticAst object
}
\description{
EinopsShapeSemanticAst constructor
}
\keyword{internal}
