% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{EllipsisAstNode}
\alias{EllipsisAstNode}
\title{Create an EllipsisAstNode}
\usage{
EllipsisAstNode(src)
}
\arguments{
\item{src}{List with start position and length}
}
\value{
EllipsisAstNode object
}
\description{
Create an EllipsisAstNode
}
\keyword{internal}
