% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R, R/utils.R
\name{execute_reduction}
\alias{execute_reduction}
\title{Execute the reduction operation}
\usage{
execute_reduction(x, reduction_plan, func, ...)

execute_reduction(x, reduction_plan, func, ...)
}
\arguments{
\item{x}{input array}

\item{reduction_plan}{list from create_reduction_plan}

\item{func}{reduction function}

\item{...}{additional arguments for func}
}
\value{
reduced array

reduced array
}
\description{
Execute the reduction operation

Execute the reduction operation
}
\keyword{internal}
