% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-parser.R
\name{find_top_level_arrow}
\alias{find_top_level_arrow}
\title{Find the top-level arrow in a token sequence}
\usage{
find_top_level_arrow(tokens)
}
\arguments{
\item{tokens}{EinopsTokenSequence object}
}
\value{
Integer position of the top-level arrow token
}
\description{
Find the top-level arrow in a token sequence
}
\keyword{internal}
