% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-parser.R
\name{generate_constructor}
\alias{generate_constructor}
\title{Generate constructor code for AST node}
\usage{
generate_constructor(node, indent_level = 0)
}
\arguments{
\item{node}{AST node}

\item{indent_level}{Integer, current indentation level (for internal use)}
}
\value{
Character string with constructor call
}
\description{
Generate constructor code for AST node
}
\keyword{internal}
