% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-backends.R
\name{get_backend}
\alias{get_backend}
\title{Main function to detect and return backend}
\usage{
get_backend(tensor)
}
\arguments{
\item{tensor}{any support tensor-like class}
}
\value{
An instance of a \code{BackendRegistry} class
}
\description{
Main function to detect and return backend
}
\keyword{internal}
