% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-ast.R
\name{NamedAxisAstNode}
\alias{NamedAxisAstNode}
\title{AST Node Constructors and Einops Parser}
\usage{
NamedAxisAstNode(name, src)
}
\arguments{
\item{name}{Character string, the name of the axis}

\item{src}{List with start position}
}
\value{
NamedAxisAstNode object
}
\description{
Functions to create AST nodes and parse einops patterns
}
\keyword{internal}
