% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{parse_einops}
\alias{parse_einops}
\title{Parse an einops expression into an EinopsAst object}
\usage{
parse_einops(expr)
}
\arguments{
\item{expr}{character string containing the einops expression}
}
\value{
EinopsAst object
}
\description{
Parse an einops expression into an EinopsAst object
}
\examples{
# Parse a simple reduction
ast <- parse_einops("h w c -> h w")

# Parse with parentheses  
ast <- parse_einops("(h w) c -> h w c")

# Parse with ellipsis
ast <- parse_einops("... h w -> ...")
}
\keyword{internal}
