% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-parser.R
\name{parse_einops_ast}
\alias{parse_einops_ast}
\title{Parse einops pattern into AST}
\usage{
parse_einops_ast(tokens)
}
\arguments{
\item{tokens}{EinopsTokenSequence object from the lexer}
}
\value{
EinopsAst object
}
\description{
Parse einops pattern into AST
}
