% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_shape.R
\name{parse_shape_tokens}
\alias{parse_shape_tokens}
\title{Parse shape tokens into AST, then does semantic Analysis}
\usage{
parse_shape_tokens(ast, dimensions)
}
\arguments{
\item{ast}{OneSidedAstNode}

\item{dimensions}{numeric vector}
}
\value{
EinopsShapeSemanticAst object
}
\description{
Parse shape tokens into AST, then does semantic Analysis
}
\keyword{internal}
