% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pprint.R
\name{pprint}
\alias{pprint}
\title{Pretty Print}
\usage{
pprint(x, ...)
}
\arguments{
\item{x}{Object to pretty print}

\item{...}{Additional arguments passed to \code{repr()}}
}
\value{
The input object, invisibly
}
\description{
This is a convenience function that prints the \code{\link[=repr]{repr()}} of an object.
It is similar to python's \code{pprint.pprint()}. Usually, to ensure that
an object is displayed in the terminal using its pprint format, just
define a \code{print.that_object_class} method that calls \code{pprint()}.
}
\keyword{internal}
