% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler-TransformRecipe.R
\name{prepare_transformation_recipe}
\alias{prepare_transformation_recipe}
\title{Create the Transformation Recipe for an einops call}
\usage{
prepare_transformation_recipe(expr, func, axes_names, ndim)
}
\arguments{
\item{expr}{The input einops expression string}

\item{func}{The string/function indicating the reduction operation}

\item{axes_names}{the user defined keyword args for dims as a \code{\link[=list]{list()}}.
Names correspond to axes variable names, and values are their dimension.}

\item{ndim}{count for the number of dimensions of the input tensor}
}
\value{
a populated \code{\link[=TransformRecipe]{TransformRecipe()}} object
}
\description{
This function does the following parts of the einops 'compilation' pipeline:
\enumerate{
\item Lexing: tokenizing the input expression string
\item Parsing: converting the tokens into an Abstract Syntax Tree (AST)
\item Syntactic Analysis:
\itemize{
\item operation-based AST validation pass
\item \link{TODO} Compile syntactic info for intermediate representation (IR).
}
\item \link{TODO} IR generation: return the \code{\link[=TransformRecipe]{TransformRecipe()}} object, acting as
the IR for the einops.
}
}
\keyword{internal}
