% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenizer.R
\name{print.EinopsTokenSequence}
\alias{print.EinopsTokenSequence}
\title{Print method for EinopsTokenSequence}
\usage{
\method{print}{EinopsTokenSequence}(x, ...)
}
\arguments{
\item{x}{EinopsTokenSequence object}

\item{...}{additional arguments (unused)}
}
\value{
invisible x
}
\description{
Print EinopsTokenSequence objects in a clean format
}
