% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pprint-repr.R
\name{repr}
\alias{repr}
\title{Python-like Representation of Objects as Strings}
\usage{
repr(x, indent = 0L, ...)
}
\arguments{
\item{x}{Object to represent}

\item{indent}{Indentation level (number of spaces)}

\item{...}{Additional arguments passed to methods}
}
\value{
A character vector of class c("repr_output", "character"), each
element is a line
}
\description{
This is an R implementation of python's \code{repr()} function.
All objects will have a default repr that uses its print method.
This is important because the \code{\link[=pprint]{pprint()}} function will always
be able to give a repr-esque output of any object.
}
\keyword{internal}
