% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R, R/tokenizer.R
\name{tokenize}
\alias{tokenize}
\title{Tokenize einops pattern into structured tokens}
\usage{
tokenize(pattern)

tokenize(pattern)
}
\arguments{
\item{pattern}{character string with einops pattern}
}
\value{
list of token objects with type, value, start, end fields

list of token objects with type, value, start, end fields
}
\description{
Tokenize einops pattern into structured tokens

Tokenize einops pattern into structured tokens
}
\keyword{internal}
